requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0

supabase==2.4.0
# 나머지 gotrue, httpx 는 pip 에 맡김
